Amended slowly moving code out of dataservice
